requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchaudio
chatterbox-tts
torchcodec
pydub